[Tools] Add a tool for using QWERTY
New page /Tools lists JavaScript tools, starting with a tool for allowing
touch-typing by people who want dvorak when the keyboard is in QWERTY, and vice
versa.

Excluded from sitemap since this is intended to be a collection of utilities
that I use but probably aren't useful for anyone and don't need to be added to
google. Not excluded from robots.txt though just in case anyone cares.