[Pages] Make routing case insensitive, use attribute for sitemap
The new `SiteMapEntry` attribute controls the paths under which a page is
included in a sitemap, instead of just always using the path from the router.
This frees up the router to use case-insensitive matching, and eliminates the
manual check to exclude tools from the sitemap.

In the process, remove the robots.txt page from the sitemap, it isn't needed.